+++ packages/python/plotly/plotly/graph_objs/layout/_shape.py
@@ -433,10 +433,7 @@ class Shape(_BaseLayoutHierarchyType):
         domain length from the left of the domain of that axis: e.g.,
         *x2 domain* refers to the domain of the second x  axis and a x
         position of 0.5 refers to the point between the left and the
-        right of the domain of the second x axis. If the axis `type` is
-        "log", then you must take the log of your desired range. If the
-        axis `type` is "date", then you must convert the date to unix
-        time in milliseconds.
+        right of the domain of the second x axis.
 
         The 'xref' property is an enumeration that may be specified as:
           - One of the following enumeration values:
@@ -550,11 +547,11 @@ class Shape(_BaseLayoutHierarchyType):
     @property
     def yref(self):
         """
-        Sets the annotation's y coordinate axis. If set to a y axis id
-        (e.g. "y" or "y2"), the `y` position refers to a y coordinate.
-        If set to "paper", the `y` position refers to the distance from
-        the bottom of the plotting area in normalized coordinates where
-        0 (1) corresponds to the bottom (top). If set to a y axis ID
+        Sets the shape's y coordinate axis. If set to a y axis id (e.g.
+        "y" or "y2"), the `y` position refers to a y coordinate. If set
+        to "paper", the `y` position refers to the distance from the
+        bottom of the plotting area in normalized coordinates where 0
+        (1) corresponds to the bottom (top). If set to a y axis ID
         followed by "domain" (separated by a space), the position
         behaves like for "paper", but refers to the distance in
         fractions of the domain length from the bottom of the domain of
@@ -714,10 +711,7 @@ class Shape(_BaseLayoutHierarchyType):
             domain of that axis: e.g., *x2 domain* refers to the
             domain of the second x  axis and a x position of 0.5
             refers to the point between the left and the right of
-            the domain of the second x axis. If the axis `type` is
-            "log", then you must take the log of your desired
-            range. If the axis `type` is "date", then you must
-            convert the date to unix time in milliseconds.
+            the domain of the second x axis.
         xsizemode
             Sets the shapes's sizing mode along the x axis. If set
             to "scaled", `x0`, `x1` and x coordinates within `path`
@@ -742,12 +736,12 @@ class Shape(_BaseLayoutHierarchyType):
             to a certain data value. No effect when `ysizemode` not
             set to "pixel".
         yref
-            Sets the annotation's y coordinate axis. If set to a y
-            axis id (e.g. "y" or "y2"), the `y` position refers to
-            a y coordinate. If set to "paper", the `y` position
-            refers to the distance from the bottom of the plotting
-            area in normalized coordinates where 0 (1) corresponds
-            to the bottom (top). If set to a y axis ID followed by
+            Sets the shape's y coordinate axis. If set to a y axis
+            id (e.g. "y" or "y2"), the `y` position refers to a y
+            coordinate. If set to "paper", the `y` position refers
+            to the distance from the bottom of the plotting area in
+            normalized coordinates where 0 (1) corresponds to the
+            bottom (top). If set to a y axis ID followed by
             "domain" (separated by a space), the position behaves
             like for "paper", but refers to the distance in
             fractions of the domain length from the bottom of the
@@ -905,10 +899,7 @@ class Shape(_BaseLayoutHierarchyType):
             domain of that axis: e.g., *x2 domain* refers to the
             domain of the second x  axis and a x position of 0.5
             refers to the point between the left and the right of
-            the domain of the second x axis. If the axis `type` is
-            "log", then you must take the log of your desired
-            range. If the axis `type` is "date", then you must
-            convert the date to unix time in milliseconds.
+            the domain of the second x axis.
         xsizemode
             Sets the shapes's sizing mode along the x axis. If set
             to "scaled", `x0`, `x1` and x coordinates within `path`
@@ -933,12 +924,12 @@ class Shape(_BaseLayoutHierarchyType):
             to a certain data value. No effect when `ysizemode` not
             set to "pixel".
         yref
-            Sets the annotation's y coordinate axis. If set to a y
-            axis id (e.g. "y" or "y2"), the `y` position refers to
-            a y coordinate. If set to "paper", the `y` position
-            refers to the distance from the bottom of the plotting
-            area in normalized coordinates where 0 (1) corresponds
-            to the bottom (top). If set to a y axis ID followed by
+            Sets the shape's y coordinate axis. If set to a y axis
+            id (e.g. "y" or "y2"), the `y` position refers to a y
+            coordinate. If set to "paper", the `y` position refers
+            to the distance from the bottom of the plotting area in
+            normalized coordinates where 0 (1) corresponds to the
+            bottom (top). If set to a y axis ID followed by
             "domain" (separated by a space), the position behaves
             like for "paper", but refers to the distance in
             fractions of the domain length from the bottom of the