+++ packages/python/plotly/plotly/validators/parcoords/__init__.py
@@ -3,6 +3,7 @@ from typing import TYPE_CHECKING
 
 if sys.version_info < (3, 7) or TYPE_CHECKING:
     from ._visible import VisibleValidator
+    from ._unselected import UnselectedValidator
     from ._uirevision import UirevisionValidator
     from ._uid import UidValidator
     from ._tickfont import TickfontValidator
@@ -32,6 +33,7 @@ else:
         [],
         [
             "._visible.VisibleValidator",
+            "._unselected.UnselectedValidator",
             "._uirevision.UirevisionValidator",
             "._uid.UidValidator",
             "._tickfont.TickfontValidator",